chore: Improve form usability with CTRL + Enter submission
The code changes in this commit add the functionality to submit the form by pressing CTRL + Enter. This improvement enhances the usability of the form, allowing users to quickly submit their input without having to click the "Translate" button.

Note: The commit message follows the established convention of using a prefix (chore) to indicate a non-functional change or maintenance task.